payment: Move payment metadata
Move shared payment metadata out of the BOLT11 module.

This lets other protocols reuse it without a BOLT11 dependency.
